tinygo: cannot call global() inside init
