+++ dom_no_tinygo.go
@@ -1,0 +1,15 @@
+// +build !tinygo
+
+package vecty
+
+func init() {
+	if isTest {
+		return
+	}
+	if global() == nil {
+		panic("vecty: only WebAssembly, TinyGo, and testing compilation is supported")
+	}
+	if global().Get("document").IsUndefined() {
+		panic("vecty: only running inside a browser is supported")
+	}
+}
